# api/database/handlers.go
package database

import (
	"context"
	_ "embed"
	"encoding/json"
	"net/http"
	"strconv"
	"strings"
)

//go:embed openapi.yaml
var openapiSpec []byte

// RegisterRoutes registers all API routes on the provided ServeMux.
//
// Routes:
//   - GET /health - Health check endpoint
//   - GET /openapi.yaml - OpenAPI 3.0 specification
//   - GET /docs - Swagger UI documentation
//   - GET/POST/PATCH/DELETE /query/{table} - CRUD operations on table rows
//   - GET /schema - List all tables in schema
//   - POST /schema - Execute raw schema SQL
//   - POST /schema/invalidate - Refresh schema cache
//   - GET/POST/DELETE/PATCH /schema/table/{table} - Table schema operations
//   - GET/POST/DELETE /schema/fts/{table} - FTS (Full-Text Search) operations
//   - GET/POST/PATCH/DELETE /db - Database management
//
// Use DB-Name header to target external Turso databases (default: primary).
func RegisterRoutes(app *http.ServeMux) {
	// Health check (no auth required)
	app.HandleFunc("GET /health", handleHealth())

	// OpenAPI documentation (no auth required)
	app.HandleFunc("GET /openapi.yaml", handleOpenAPISpec())
	app.HandleFunc("GET /docs", handleSwaggerUI())

	// Row operations
	app.HandleFunc("GET /query/{table}", handleSelectRows())
	app.HandleFunc("POST /query/{table}", handleInsertRows())
	app.HandleFunc("PATCH /query/{table}", handleUpdateRows())
	app.HandleFunc("DELETE /query/{table}", handleDeleteRows())

	// Schema operations
	app.HandleFunc("GET /schema", handleGetSchema())
	app.HandleFunc("POST /schema", handleEditSchema())
	app.HandleFunc("POST /schema/invalidate", handleInvalidateSchema())

	// Table schema operations
	app.HandleFunc("GET /schema/table/{table}", handleGetTableSchema())
	app.HandleFunc("POST /schema/table/{table}", handleCreateTable())
	app.HandleFunc("DELETE /schema/table/{table}", handleDropTable())
	app.HandleFunc("PATCH /schema/table/{table}", handleAlterTable())

	// FTS (Full-Text Search) operations
	app.HandleFunc("GET /schema/fts", handleListFTSIndexes())
	app.HandleFunc("POST /schema/fts/{table}", handleCreateFTSIndex())
	app.HandleFunc("DELETE /schema/fts/{table}", handleDropFTSIndex())

	// Database management
	app.HandleFunc("GET /db", handleListDbs())
	app.HandleFunc("POST /db", handleCreateDb())
	app.HandleFunc("PATCH /db", handleRegisterDb())
	app.HandleFunc("PATCH /db/all", handleRegisterAll())
	app.HandleFunc("DELETE /db/{name}", handleDeleteDb())

	// Schema template management
	app.HandleFunc("GET /templates", handleListTemplates())
	app.HandleFunc("POST /templates", handleCreateTemplate())
	app.HandleFunc("GET /templates/{name}", handleGetTemplate())
	app.HandleFunc("PUT /templates/{name}", handleUpdateTemplate())
	app.HandleFunc("DELETE /templates/{name}", handleDeleteTemplate())
	app.HandleFunc("POST /templates/{name}/sync", handleSyncTemplate())
	app.HandleFunc("GET /templates/{name}/databases", handleListTemplateDBs())

	// Database-template association
	app.HandleFunc("GET /db/{name}/template", handleGetDBTemplate())
	app.HandleFunc("PUT /db/{name}/template", handleSetDBTemplate())
	app.HandleFunc("DELETE /db/{name}/template", handleRemoveDBTemplate())
	app.HandleFunc("POST /db/{name}/sync", handleSyncDBToTemplate())
}

func handleSelectRows() http.HandlerFunc {
	return withDBResponse(func(ctx context.Context, dao *Database, req *http.Request, w http.ResponseWriter) ([]byte, error) {
		params := req.URL.Query()

		// Check for count preferences
		prefer := req.Header.Get("Prefer")
		includeCount := strings.Contains(prefer, "count=exact")
		countOnly := params.Get(ParamCount) == "only"

		// If count is requested, use SelectWithCount
		if includeCount || countOnly {
			result, err := dao.SelectWithCount(ctx, req.PathValue("table"), params, includeCount, countOnly)
			if err != nil {
				return nil, err
			}

			// Set count header
			if includeCount || countOnly {
				w.Header().Set("X-Total-Count", strconv.FormatInt(result.Count, 10))
			}

			// If count only, return just the count as JSON
			if countOnly {
				return json.Marshal(map[string]int64{"count": result.Count})
			}

			return result.Data, nil
		}

		return dao.Select(ctx, req.PathValue("table"), params)
	})
}

func handleInsertRows() http.HandlerFunc {
	return withDB(func(ctx context.Context, dao *Database, req *http.Request) ([]byte, error) {
		if req.Header.Get("Prefer") == "resolution=merge-duplicates" {
			return dao.Upsert(ctx, req.PathValue("table"), req.URL.Query(), req.Body)
		}

		return dao.Insert(ctx, req.PathValue("table"), req.URL.Query(), req.Body)
	})
}

func handleUpdateRows() http.HandlerFunc {
	return withDB(func(ctx context.Context, dao *Database, req *http.Request) ([]byte, error) {
		return dao.Update(ctx, req.PathValue("table"), req.URL.Query(), req.Body)
	})
}

func handleDeleteRows() http.HandlerFunc {
	return withDB(func(ctx context.Context, dao *Database, req *http.Request) ([]byte, error) {
		return dao.Delete(ctx, req.PathValue("table"), req.URL.Query())
	})
}

func handleEditSchema() http.HandlerFunc {
	return withDB(func(ctx context.Context, dao *Database, req *http.Request) ([]byte, error) {
		return dao.EditSchema(ctx, req.Body)
	})
}

func handleInvalidateSchema() http.HandlerFunc {
	return withDB(func(ctx context.Context, dao *Database, req *http.Request) ([]byte, error) {
		return []byte(`{"message":"schema invalidated"}`), dao.InvalidateSchema(ctx)
	})
}

func handleGetSchema() http.HandlerFunc {
	return withDB(func(ctx context.Context, dao *Database, req *http.Request) ([]byte, error) {
		return dao.GetSchema()
	})
}

func handleGetTableSchema() http.HandlerFunc {
	return withDB(func(ctx context.Context, dao *Database, req *http.Request) ([]byte, error) {
		return dao.GetTableSchema(req.PathValue("table"))
	})
}

func handleCreateTable() http.HandlerFunc {
	return withDB(func(ctx context.Context, dao *Database, req *http.Request) ([]byte, error) {
		return dao.CreateTable(ctx, req.PathValue("table"), req.Body)
	})
}

func handleDropTable() http.HandlerFunc {
	return withDB(func(ctx context.Context, dao *Database, req *http.Request) ([]byte, error) {
		return dao.DropTable(ctx, req.PathValue("table"))
	})
}

func handleAlterTable() http.HandlerFunc {
	return withDB(func(ctx context.Context, dao *Database, req *http.Request) ([]byte, error) {
		return dao.AlterTable(ctx, req.PathValue("table"), req.Body)
	})
}

func handleListFTSIndexes() http.HandlerFunc {
	return withDB(func(ctx context.Context, dao *Database, req *http.Request) ([]byte, error) {
		return dao.ListFTSIndexes()
	})
}

func handleCreateFTSIndex() http.HandlerFunc {
	return withDB(func(ctx context.Context, dao *Database, req *http.Request) ([]byte, error) {
		return dao.CreateFTSIndex(ctx, req.PathValue("table"), req.Body)
	})
}

func handleDropFTSIndex() http.HandlerFunc {
	return withDB(func(ctx context.Context, dao *Database, req *http.Request) ([]byte, error) {
		return dao.DropFTSIndex(ctx, req.PathValue("table"))
	})
}

func handleCreateDb() http.HandlerFunc {
	return withPrimary(func(ctx context.Context, dao *PrimaryDao, req *http.Request) ([]byte, error) {
		return dao.CreateDB(ctx, req.Body)
	})
}

func handleRegisterAll() http.HandlerFunc {
	return withPrimary(func(ctx context.Context, dao *PrimaryDao, req *http.Request) ([]byte, error) {
		err := dao.RegisterAllDbs(ctx)
		return nil, err
	})
}

func handleRegisterDb() http.HandlerFunc {
	return withPrimary(func(ctx context.Context, dao *PrimaryDao, req *http.Request) ([]byte, error) {
		return dao.RegisterDB(ctx, req.Body, req.Header.Get("DB-Token"))
	})
}

func handleListDbs() http.HandlerFunc {
	return withPrimary(func(ctx context.Context, dao *PrimaryDao, req *http.Request) ([]byte, error) {
		return dao.ListDBs(ctx)
	})
}

func handleDeleteDb() http.HandlerFunc {
	return withPrimary(func(ctx context.Context, dao *PrimaryDao, req *http.Request) ([]byte, error) {
		return dao.DeleteDB(ctx, req.PathValue("name"))
	})
}

func handleHealth() http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		// Check database connectivity using the connection pool
		dao, err := ConnPrimary()
		if err != nil {
			w.Header().Set("Content-Type", "application/json")
			w.WriteHeader(http.StatusServiceUnavailable)
			w.Write([]byte(`{"status":"unhealthy","error":"database connection failed"}`))
			return
		}
		// Note: don't close dao.Client - it's managed by the connection pool

		if err := dao.Client.PingContext(r.Context()); err != nil {
			w.Header().Set("Content-Type", "application/json")
			w.WriteHeader(http.StatusServiceUnavailable)
			w.Write([]byte(`{"status":"unhealthy","error":"database ping failed"}`))
			return
		}

		w.Header().Set("Content-Type", "application/json")
		w.WriteHeader(http.StatusOK)
		w.Write([]byte(`{"status":"healthy"}`))
	}
}

func handleOpenAPISpec() http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		w.Header().Set("Content-Type", "application/x-yaml")
		w.Header().Set("Access-Control-Allow-Origin", "*")
		w.Write(openapiSpec)
	}
}

func handleSwaggerUI() http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		w.Header().Set("Content-Type", "text/html")
		w.Write([]byte(`<!DOCTYPE html>
<html>
<head>
  <title>Atomicbase API</title>
  <link rel="stylesheet" href="https://unpkg.com/swagger-ui-dist@5/swagger-ui.css">
</head>
<body>
  <div id="swagger-ui"></div>
  <script src="https://unpkg.com/swagger-ui-dist@5/swagger-ui-bundle.js"></script>
  <script>
    SwaggerUIBundle({
      url: '/openapi.yaml',
      dom_id: '#swagger-ui',
      presets: [SwaggerUIBundle.presets.apis, SwaggerUIBundle.SwaggerUIStandalonePreset],
      layout: 'BaseLayout'
    });
  </script>
</body>
</html>`))
	}
}

// Template handlers

func handleListTemplates() http.HandlerFunc {
	return withPrimary(func(ctx context.Context, dao *PrimaryDao, req *http.Request) ([]byte, error) {
		return dao.ListTemplatesJSON(ctx)
	})
}

func handleCreateTemplate() http.HandlerFunc {
	return withPrimary(func(ctx context.Context, dao *PrimaryDao, req *http.Request) ([]byte, error) {
		type reqBody struct {
			Name   string  `json:"name"`
			Tables []Table `json:"tables"`
		}
		var body reqBody
		if err := json.NewDecoder(req.Body).Decode(&body); err != nil {
			return nil, err
		}
		template, err := dao.CreateTemplate(ctx, body.Name, body.Tables)
		if err != nil {
			return nil, err
		}
		return json.Marshal(template)
	})
}

func handleGetTemplate() http.HandlerFunc {
	return withPrimary(func(ctx context.Context, dao *PrimaryDao, req *http.Request) ([]byte, error) {
		template, err := dao.GetTemplate(ctx, req.PathValue("name"))
		if err != nil {
			return nil, err
		}
		return json.Marshal(template)
	})
}

func handleUpdateTemplate() http.HandlerFunc {
	return withPrimary(func(ctx context.Context, dao *PrimaryDao, req *http.Request) ([]byte, error) {
		type reqBody struct {
			Tables []Table `json:"tables"`
		}
		var body reqBody
		if err := json.NewDecoder(req.Body).Decode(&body); err != nil {
			return nil, err
		}
		template, err := dao.UpdateTemplate(ctx, req.PathValue("name"), body.Tables)
		if err != nil {
			return nil, err
		}
		return json.Marshal(template)
	})
}

func handleDeleteTemplate() http.HandlerFunc {
	return withPrimary(func(ctx context.Context, dao *PrimaryDao, req *http.Request) ([]byte, error) {
		if err := dao.DeleteTemplate(ctx, req.PathValue("name")); err != nil {
			return nil, err
		}
		return []byte(`{"message":"template deleted"}`), nil
	})
}

func handleSyncTemplate() http.HandlerFunc {
	return withPrimary(func(ctx context.Context, dao *PrimaryDao, req *http.Request) ([]byte, error) {
		dropExtra := req.URL.Query().Get("dropExtra") == "true"
		results, err := dao.SyncTemplate(ctx, req.PathValue("name"), dropExtra)
		if err != nil {
			return nil, err
		}
		return json.Marshal(results)
	})
}

func handleListTemplateDBs() http.HandlerFunc {
	return withPrimary(func(ctx context.Context, dao *PrimaryDao, req *http.Request) ([]byte, error) {
		databases, err := dao.ListDatabasesByTemplate(ctx, req.PathValue("name"))
		if err != nil {
			return nil, err
		}
		if databases == nil {
			databases = []string{}
		}
		return json.Marshal(databases)
	})
}

func handleGetDBTemplate() http.HandlerFunc {
	return withPrimary(func(ctx context.Context, dao *PrimaryDao, req *http.Request) ([]byte, error) {
		template, err := dao.GetDatabaseTemplate(ctx, req.PathValue("name"))
		if err != nil {
			return nil, err
		}
		if template == nil {
			return []byte("null"), nil
		}
		return json.Marshal(template)
	})
}

func handleSetDBTemplate() http.HandlerFunc {
	return withPrimary(func(ctx context.Context, dao *PrimaryDao, req *http.Request) ([]byte, error) {
		type reqBody struct {
			TemplateName string `json:"templateName"`
		}
		var body reqBody
		if err := json.NewDecoder(req.Body).Decode(&body); err != nil {
			return nil, err
		}
		if err := dao.AssociateTemplate(ctx, req.PathValue("name"), body.TemplateName); err != nil {
			return nil, err
		}
		return []byte(`{"message":"template associated"}`), nil
	})
}

func handleRemoveDBTemplate() http.HandlerFunc {
	return withPrimary(func(ctx context.Context, dao *PrimaryDao, req *http.Request) ([]byte, error) {
		if err := dao.DisassociateTemplate(ctx, req.PathValue("name")); err != nil {
			return nil, err
		}
		return []byte(`{"message":"template disassociated"}`), nil
	})
}

func handleSyncDBToTemplate() http.HandlerFunc {
	return withPrimary(func(ctx context.Context, dao *PrimaryDao, req *http.Request) ([]byte, error) {
		dropExtra := req.URL.Query().Get("dropExtra") == "true"
		changes, err := dao.SyncDatabaseToTemplate(ctx, req.PathValue("name"), dropExtra)
		if err != nil {
			return nil, err
		}
		if changes == nil {
			changes = []string{}
		}
		return json.Marshal(map[string]any{
			"success": true,
			"changes": changes,
		})
	})
}