# api/database/openapi.yaml
openapi: 3.0.3
info:
  title: Atomicbase API
  description: |
    PostgREST-style REST API for SQLite and Turso databases.

    ## Authentication
    If `ATOMICBASE_API_KEY` is set, all endpoints (except `/health`) require a Bearer token:
    ```
    Authorization: Bearer <your-api-key>
    ```

    ## Targeting Databases
    By default, requests target the primary SQLite database. To target an external Turso database, use the `DB-Name` header:
    ```
    DB-Name: my-turso-db
    ```

    ## Query Syntax

    ### Filtering
    Use query parameters with operators:
    - `column=eq.value` - Equal to
    - `column=neq.value` - Not equal to
    - `column=gt.value` - Greater than
    - `column=gte.value` - Greater than or equal
    - `column=lt.value` - Less than
    - `column=lte.value` - Less than or equal
    - `column=like.pattern` - SQL LIKE pattern
    - `column=ilike.pattern` - Case-insensitive LIKE
    - `column=in.(a,b,c)` - In list
    - `column=is.null` / `column=is.true` / `column=is.false`
    - `column=fts.search terms` - Full-text search (requires FTS index)

    ### Negation (NOT operator)
    Prefix any operator with `not.` to negate it:
    - `column=not.eq.value` - Not equal to
    - `column=not.like.pattern` - Not like
    - `column=not.in.(a,b,c)` - Not in list
    - `column=not.is.null` - Is not null
    - `column=not.gt.value` - Not greater than

    ### Column Selection
    Use `select` parameter to choose columns and embed related tables:
...

...

...

...

...

...
  license:
...
servers:
...
paths:
...
  /query/{table}:
    get:
      tags:
        - Query
      summary: Select rows
      description: |
        Query rows from a table with optional filtering, ordering, pagination, and nested relations.

        Use query parameters for WHERE conditions (e.g., `id=eq.1`).

        **Aggregations**: Use aggregate functions in select parameter:
        - `select=count(*)` - Count all rows
        - `select=category,sum(price)` - Group by with aggregates

        **Row Count**: Add `Prefer: count=exact` header to include total count in `X-Total-Count` response header.
        Use `count=only` parameter to get only the count without data.
      operationId: selectRows
      parameters:
        - $ref: '#/components/parameters/TablePath'
        - $ref: '#/components/parameters/DBNameHeader'
        - $ref: '#/components/parameters/SelectParam'
        - $ref: '#/components/parameters/OrderParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
        - $ref: '#/components/parameters/OrParam'
        - $ref: '#/components/parameters/CountParam'
        - name: Prefer
          in: header
          description: Set to `count=exact` to include total row count in X-Total-Count header
          schema:
            type: string
            enum:
              - count=exact
      responses:
        '200':
          description: Rows matching the query
          headers:
            X-Total-Count:
              description: Total row count (when Prefer count=exact is set)
              schema:
                type: integer
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      type: object
                  - $ref: '#/components/schemas/CountResponse'
              examples:
                rows:
                  summary: Normal row response
                  value:
                    - id: 1
                      name: "Alice"
                      email: "alice@example.com"
                    - id: 2
                      name: "Bob"
                      email: "bob@example.com"
                countOnly:
                  summary: Count only response
                  value:
                    count: 42
                aggregates:
                  summary: Aggregation response
                  value:
                    - category: "electronics"
                      count: 15
                      sum_price: 4500.00
                    - category: "books"
                      count: 30
                      sum_price: 750.50
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
      security:
        - bearerAuth: []

    post:
      tags:
        - Query
      summary: Insert row
      description: |
        Insert a single row into a table.

        To perform an upsert (insert or update on conflict), add the header:
        ```
        Prefer: resolution=merge-duplicates
        ```
        For upsert, the request body should be an array of objects.
      operationId: insertRow
      parameters:
        - $ref: '#/components/parameters/TablePath'
        - $ref: '#/components/parameters/DBNameHeader'
        - $ref: '#/components/parameters/SelectParam'
        - name: Prefer
          in: header
          description: Set to `resolution=merge-duplicates` for upsert behavior
          schema:
            type: string
            enum:
              - resolution=merge-duplicates
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                  description: Single row insert
                  additionalProperties: true
                - type: array
                  description: Multiple rows for upsert
                  items:
                    type: object
                    additionalProperties: true
            examples:
              insert:
                summary: Single row insert
                value:
                  name: "Charlie"
                  email: "charlie@example.com"
              upsert:
                summary: Upsert multiple rows
                value:
                  - id: 1
                    name: "Alice Updated"
                  - id: 3
                    name: "New User"
      responses:
        '200':
          description: Row inserted successfully
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/InsertResponse'
                  - $ref: '#/components/schemas/RowsAffectedResponse'
                  - type: array
                    items:
                      type: object
              examples:
                insert:
                  summary: Insert without select
                  value:
                    last_insert_id: 3
                upsert:
                  summary: Upsert without select
                  value:
                    rows_affected: 2
                withSelect:
                  summary: With select parameter
                  value:
                    - id: 3
                      name: "Charlie"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Conflict (unique constraint violation)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: record already exists
        '429':
          $ref: '#/components/responses/RateLimited'
      security:
        - bearerAuth: []

    patch:
      tags:
        - Query
      summary: Update rows
      description: |
        Update rows matching the WHERE conditions from query parameters.

        Use query parameters for WHERE conditions (e.g., `id=eq.1`).
      operationId: updateRows
      parameters:
        - $ref: '#/components/parameters/TablePath'
        - $ref: '#/components/parameters/DBNameHeader'
        - $ref: '#/components/parameters/SelectParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              description: Column values to update
            example:
              name: "Alice Updated"
              email: "alice.new@example.com"
      responses:
        '200':
          description: Rows updated successfully
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/RowsAffectedResponse'
                  - type: array
                    items:
                      type: object
              examples:
                default:
                  summary: Without select parameter
                  value:
                    rows_affected: 1
                withSelect:
                  summary: With select parameter
                  value:
                    - id: 1
                      name: "Alice Updated"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
      security:
        - bearerAuth: []

    delete:
      tags:
        - Query
      summary: Delete rows
      description: |
        Delete rows matching the WHERE conditions from query parameters.

        **A WHERE clause is required** - mass deletes without conditions are not allowed.
      operationId: deleteRows
      parameters:
        - $ref: '#/components/parameters/TablePath'
        - $ref: '#/components/parameters/DBNameHeader'
        - $ref: '#/components/parameters/SelectParam'
      responses:
        '200':
          description: Rows deleted successfully
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/RowsAffectedResponse'
                  - type: array
                    items:
                      type: object
              examples:
                default:
                  summary: Without select parameter
                  value:
                    rows_affected: 1
                withSelect:
                  summary: With select parameter (returns deleted rows)
                  value:
                    - id: 1
                      name: "Deleted User"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
      security:
        - bearerAuth: []
  /schema:
    get:
      tags:
        - Schema
      summary: List all tables
      description: |
        Get a list of all tables in the database schema with their columns and primary keys.
      operationId: getSchema
      parameters:
        - $ref: '#/components/parameters/DBNameHeader'
      responses:
        '200':
          description: List of tables
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Table'
              example:
                - name: users
                  pk: id
                  columns:
                    - name: id
                      type: INTEGER
                    - name: email
                      type: TEXT
                    - name: name
                      type: TEXT
                - name: posts
                  pk: id
                  columns:
                    - name: id
                      type: INTEGER
                    - name: title
                      type: TEXT
                    - name: user_id
                      type: INTEGER
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
      security:
        - bearerAuth: []
    post:
      tags:
        - Schema
      summary: Execute DDL query
      description: |
        Execute a raw DDL (Data Definition Language) SQL query.

        **Only DDL statements are allowed** (CREATE, ALTER, DROP, etc.). DML statements (SELECT, INSERT, UPDATE, DELETE) will be rejected.
      operationId: executeSchema
      parameters:
        - $ref: '#/components/parameters/DBNameHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchemaQueryRequest'
            example:
              query: "CREATE INDEX idx_users_email ON users(email)"
              args: []
      responses:
        '200':
          description: Schema modified successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: schema edited
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
      security:
        - bearerAuth: []
  /schema/invalidate:
    post:
      tags:
        - Schema
      summary: Invalidate schema cache
      description: |
        Force a refresh of the schema cache.

        Call this after making schema changes outside of Atomicbase or if the cache becomes stale.
      operationId: invalidateSchema
      parameters:
        - $ref: '#/components/parameters/DBNameHeader'
      responses:
        '200':
          description: Schema cache invalidated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: schema invalidated
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
      security:
        - bearerAuth: []
  /schema/table/{table}:
    get:
      tags:
        - Schema
      summary: Get table schema
      description: Get the schema definition for a specific table including columns, primary key, and foreign keys.
      operationId: getTableSchema
      parameters:
        - $ref: '#/components/parameters/TablePath'
        - $ref: '#/components/parameters/DBNameHeader'
      responses:
        '200':
          description: Table schema
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableSchemaResponse'
              example:
                columns:
                  - name: id
                    type: INTEGER
                  - name: name
                    type: TEXT
                  - name: email
                    type: TEXT
                  - name: created_at
                    type: TEXT
                primaryKey: id
                foreignKeys:
                  - column: org_id
                    references: organizations.id
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
      security:
        - bearerAuth: []

    post:
      tags:
        - Schema
      summary: Create table
      description: Create a new table with the specified columns and constraints.
      operationId: createTable
      parameters:
        - $ref: '#/components/parameters/TablePath'
        - $ref: '#/components/parameters/DBNameHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTableRequest'
            example:
              id:
                type: integer
                primaryKey: true
              name:
                type: text
                notNull: true
              email:
                type: text
                unique: true
              org_id:
                type: integer
                references: organizations.id
                onDelete: cascade
      responses:
        '200':
          description: Table created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: table users created
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
      security:
        - bearerAuth: []

    patch:
      tags:
        - Schema
      summary: Alter table
      description: Modify an existing table structure (add columns, rename columns, drop columns, rename table).
      operationId: alterTable
      parameters:
        - $ref: '#/components/parameters/TablePath'
        - $ref: '#/components/parameters/DBNameHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlterTableRequest'
            examples:
              addColumn:
                summary: Add a new column
                value:
                  newColumns:
                    phone:
                      type: text
                      notNull: false
              renameColumn:
                summary: Rename a column
                value:
                  renameColumns:
                    old_name: new_name
              dropColumn:
                summary: Drop a column
                value:
                  dropColumns:
                    - temp_column
              renameTable:
                summary: Rename the table
                value:
                  newName: users_v2
      responses:
        '200':
          description: Table altered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: table users altered
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
      security:
        - bearerAuth: []

    delete:
      tags:
        - Schema
      summary: Drop table
      description: Drop (delete) a table and all its data.
      operationId: dropTable
      parameters:
        - $ref: '#/components/parameters/TablePath'
        - $ref: '#/components/parameters/DBNameHeader'
      responses:
        '200':
          description: Table dropped successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: table users dropped
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
      security:
        - bearerAuth: []
  /schema/fts:
    get:
      tags:
        - Schema
      summary: List FTS indexes
      description: |
        List all tables that have FTS5 (Full-Text Search) indexes configured.

        Returns information about each FTS index including the base table name, FTS table name, and indexed columns.
      operationId: listFTSIndexes
      parameters:
        - $ref: '#/components/parameters/DBNameHeader'
      responses:
        '200':
          description: List of FTS indexes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FTSIndexInfo'
              example:
                - table: articles
                  ftsTable: articles_fts
                  columns:
                    - title
                    - content
                - table: products
                  ftsTable: products_fts
                  columns:
                    - name
                    - description
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
      security:
        - bearerAuth: []
...
  /db/all:
    patch:
      tags:
        - Database
      summary: Register all databases
      description: |
        Discover and register all Turso databases in your organization.

        Requires `TURSO_ORGANIZATION` and `TURSO_API_KEY` environment variables.
      operationId: registerAllDatabases
      responses:
        '200':
          description: All databases registered successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
      security:
        - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        API key authentication. Set `ATOMICBASE_API_KEY` environment variable to enable.

        If not set, authentication is disabled.
  parameters:
    TablePath:
      name: table
      in: path
      required: true
      description: Table name
      schema:
        type: string
        pattern: '^[a-zA-Z_][a-zA-Z0-9_]*$'
        maxLength: 128
    DBNameHeader:
      name: DB-Name
      in: header
      description: Target database name (default is primary database)
      schema:
        type: string
...
    OrderParam:
      name: order
      in: query
      description: |
        Sort order for results.

        Examples:
        - `name:asc` - Ascending by name
        - `created_at:desc` - Descending by created_at
      schema:
        type: string
...